Move existing reader to V1 directory
