+++ src/Sentry.Android.AssemblyReader/AndroidAssemblyReaderFactory.cs
@@ -1,3 +1,5 @@
+using Sentry.Android.AssemblyReader.V1;
+
 namespace Sentry.Android.AssemblyReader;
 
 /// <summary>
