+++ src/Sentry.Android.AssemblyReader/V1/ATTRIBUTION.txt
