+++ src/Sentry.Android.AssemblyReader/V1/AndroidAssemblyDirectoryReader.cs
@@ -1,4 +1,4 @@
-namespace Sentry.Android.AssemblyReader;
+namespace Sentry.Android.AssemblyReader.V1;
 
 // The "Old" app type - where each DLL is placed in the 'assemblies' directory as an individual file.
 internal sealed class AndroidAssemblyDirectoryReader : AndroidAssemblyReader, IAndroidAssemblyReader
